requirements/base.txt
Django==1.7
amqp==1.4.6
anyjson==0.3.3
billiard==3.3.0.18
celery==3.1.13
django-redis==3.7.1
django-reversion==1.8.2
django-selectable==0.8.0
django-taggit==0.12.2
django-treebeard==2.0
djangorestframework==2.3.14
docutils==0.12
kombu==3.0.21
lxml==3.3.5
pillow==2.5.3
psycopg2==2.5.3
python-docx==0.7.4
python-pptx==0.4.2
pytz==2014.4
redis==2.10.3
requests==2.3.0
six==1.7.3
sqlparse==0.1.11
unicodecsv==0.9.4
xlrd==0.9.3
xlwt==0.7.5
